Add optional "settingsPath" and "domPersistencePath" settings to SharpDevelop.exe.config. These can be used to redirect the settings/code completion cache to different folders, e.g. for running SharpDevelop from a USB stick. Make XML documentation tooltips work even if the code completion cache is disabled.
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@2596 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Main/ICSharpCode.SharpDevelop.Dom/Project/Src/XmlDoc.cs
@@ -240,20 +240,23 @@ namespace ICSharpCode.SharpDevelop.Dom
 		public static XmlDoc Load(string fileName, string cachePath)
 		{
 			//LoggingService.Debug("Loading XmlDoc for " + fileName);
-			Directory.CreateDirectory(cachePath);
-			string cacheName = cachePath + "/" + Path.GetFileNameWithoutExtension(fileName)
-				+ "." + fileName.GetHashCode().ToString("x") + ".dat";
 			XmlDoc doc;
-			if (File.Exists(cacheName)) {
-				doc = new XmlDoc();
-				if (doc.LoadFromBinary(cacheName, File.GetLastWriteTimeUtc(fileName))) {
-					//LoggingService.Debug("XmlDoc: Load from cache successful");
-					return doc;
-				} else {
-					doc.Dispose();
-					try {
-						File.Delete(cacheName);
-					} catch {}
+			string cacheName = null;
+			if (cachePath != null) {
+				Directory.CreateDirectory(cachePath);
+				cacheName = cachePath + "/" + Path.GetFileNameWithoutExtension(fileName)
+					+ "." + fileName.GetHashCode().ToString("x") + ".dat";
+				if (File.Exists(cacheName)) {
+					doc = new XmlDoc();
+					if (doc.LoadFromBinary(cacheName, File.GetLastWriteTimeUtc(fileName))) {
+						//LoggingService.Debug("XmlDoc: Load from cache successful");
+						return doc;
+					} else {
+						doc.Dispose();
+						try {
+							File.Delete(cacheName);
+						} catch {}
+					}
 				}
 			}
 			
@@ -266,7 +269,7 @@ namespace ICSharpCode.SharpDevelop.Dom
 				return new XmlDoc();
 			}
 			
-			if (doc.xmlDescription.Count > cacheLength * 2) {
+			if (cachePath != null && doc.xmlDescription.Count > cacheLength * 2) {
 				LoggingService.Debug("XmlDoc: Creating cache for " + fileName);
 				DateTime date = File.GetLastWriteTimeUtc(fileName);
 				try {
